+++ mail_private/doc/changelog.rst
@@ -1,3 +1,9 @@
+`1.1.0`
+-------
+
+- **New**: added ability to select channels for private message sending.
+- **New**: internal users are flagged automatically.
+
 `1.0.1`
 -------